Make sure restart cmd line has no duplicate parts
